aoe: allow user to disable target failure timeout
With this change, the aoe driver treats the value zero as special for the
aoe_deadsecs module parameter.  Normally, this value specifies the number
of seconds during which the driver will continue to attempt retransmits to
an unresponsive AoE target.  After aoe_deadsecs has elapsed, the aoe
driver marks the aoe device as "down" and fails all I/O.

The new meaning of an aoe_deadsecs of zero is for the driver to retransmit
commands indefinitely.

Signed-off-by: Ed Cashin <ecashin@coraid.com>
Cc: Jens Axboe <axboe@kernel.dk>
Signed-off-by: Andrew Morton <akpm@linux-foundation.org>